Renamed logo.tsx to MainLogo.tsx
