+++ src/components/MainLogo.tsx
